Removed m_fPosX,Y,and Z public vars.  Also cleaned up the links to not use them.  Had to go through the plugins to remove references to them, and I was pretty careful.
Fixed some issues in the ode plugin with virtual inheritance.  ALso fixed a bug in the generic body where the position was getting set in the base init only in the base interface, not on the body.  This was causing some bad problems with inertia in ODE if the body was starting any real distance from the origin.

Also put some code in the ode plugin to fallback gracefully if calculating the inertia on any geometry fails. I seem to get negative inertia values on some meshes.  I'm not sure why.

the novodex and bullet plugins now call the superclass init in generic body.  They were replicating some code.

git-svn-id: https://svn.code.sf.net/p/pal/code@233 62ce3681-6fdb-4e3c-a3e8-19904b470fb8