Change the pal vehicle significantly.  It now has a struct for the wheel config so paramaters can be added more easily.  Mode the roll influence a public parameter because configuring the front and real roll center is a common necessary option.
The bullet version also takes into account an odd behavior where it multiplies the forces of the suspension by the mass.  To fix this, it now divides out the mass before setting these parameters.  It also sets the max suspension force now to be 3 times the curb load of each tire.  This makes it so you don't have to set this publically, but you get a decent behavior without having to set the value.  For now, I chose not to make it public because having a max suspension force doesn't really make sense from a external perspective.

The netwon code has been update to use the new API. 

Also made get location matrix for bullet use the motion state if it has one so that you get the interpolation when running bullet.


git-svn-id: https://svn.code.sf.net/p/pal/code@429 62ce3681-6fdb-4e3c-a3e8-19904b470fb8